requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
pydantic>=2.7.0
httpx>=0.27.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
